Change TotalAmount type from int to decimal in Invoice class

+++ src/PaymentService.Domain/Entities/Invoice.cs
@@ -5,7 +5,7 @@ public class Invoice
     int Id;
     int BookingId;
     InvoiceStatus Status { get; set; } = InvoiceStatus.Pending;
-    int TotalAmount;
+    decimal TotalAmount;
     List<InvoiceLine> Lines;
     DateTime CreatedAt;
     DateTime UpdatedAt;
